backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
watchdog==3.0.0
# TensorFlow is optional - install separately if needed for advanced audio classification
# tensorflow==2.15.0
# tensorflow-hub==0.15.0
scenedetect[opencv]==0.6.2
opentimelineio==0.15.0
openai-whisper==20231117
ffmpeg-python==0.2.0
openai>=1.3.5
langchain>=0.0.350
pydantic==2.5.0
python-dotenv==1.0.0
numpy==1.24.3
librosa==0.10.1
soundfile==0.12.1
aiofiles==23.2.1
pydub==0.25.1
pandas==2.0.3
opencv-python==4.8.1.78
typing-extensions==4.8.0

